Started dealing cards to players
Apparently, when one instance of a class is multiplied in python, it just makes multiple pointers to the same thing, which is why the player name was the same for all players. Had to use list comprehension for that.
The implementation of deal was incorrect. It was horribly made for some reason. The implementation made no sense, so, had to change that.